+++ src/sampling.h
@@ -1,0 +1,10 @@
+#include <R.h>
+#include <Rmath.h>
+
+#ifndef _SAMPLING_H
+#define _SAMPLING_H
+
+// prototypes for the functions
+void pps_prob(double*, int*, int*, double*);
+void pps_brewer(double*, int*, int*, int*);
+#endif